udemy-steven-grieder/facts-reader/requirements.txt
langchain
langchain-openai
langchain-community
openai
python-dotenv
chromadb